Fix backend CI shard failures
